code/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
astroid==3.3.8
certifi==2024.8.30
dill==0.3.9
distro==1.9.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
isort==6.0.1
jiter==0.7.0
mccabe==0.7.0
openai==1.53.0
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
psutil==7.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pylint==3.3.4
pytest==8.3.5
python-dotenv==1.0.1
sniffio==1.3.1
tomlkit==0.13.2
tqdm==4.66.6
typing_extensions==4.12.2
